Get the test suite running using the new testing support library
