+++ squidb-tests/AndroidManifest.xml
@@ -10,26 +10,8 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:versionCode="1"
     android:versionName="1.0">
-
     <uses-sdk
         android:minSdkVersion="14"
         android:targetSdkVersion="24"
         tools:overrideLibrary="com.yahoo.squidb.sqlitebindings" />
-
-    <instrumentation
-        android:name="com.yahoo.squidb.android.SquidInstrumentationTestRunner"
-        android:targetPackage="com.yahoo.squidb.test" />
-
-    <application
-        android:icon="@drawable/ic_launcher"
-        android:label="@string/app_name">
-        <uses-library android:name="android.test.runner" />
-
-        <activity android:name="com.yahoo.squidb.android.TestActivity">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-            </intent-filter>
-        </activity>
-    </application>
-
 </manifest>
